Here is a possible commit message for the staged changes:
 "Add border option to diagnostic float (#647)"

This commit message describes the changes made in the `lua/plugins/lsp.lua` file and provides some context for the changes. The message includes information about the specific issue or feature being addressed by the commit, in this case, adding a "single" border option to the diagnostic float. The message also includes the issue number (#647) for reference purposes.