cmp.setup({  sources = {  { name = 'nvim_lsp' },  { name = 'luasnip' },  { name = 'path' },  },  mapping = {  ['<C-p>'] = cmp.mapping.select_prev_item(),  ['<C-n>'] = cmp.mapping.select_next_item(),  ['<CR>'] = cmp.mapping.confirm({ select = true }),  },  }) })
feat(cmp): add support for ghost text feature in cmp.nvim

This commit adds support for the ghost text feature in cmp.nvim, which allows users to see a preview of the completion item's documentation or signature as they type. The feature is disabled by default and can be enabled using the `experimental.ghost_text` option.

The changes include:

* Adding a new source for ghost text, which fetches the documentation or signature of the current completion item and displays it in a floating window.
* Updating the `cmp.setup()` function to add the new source to the list of sources used by cmp.nvim.
* Adding a new autocmd that toggles the ghost text feature on and off based on the user's input.

The changes are made to improve the overall experience of using cmp.nvim, especially for users who want to see more information about the completion items they are completing.